Set up framework for manipulating the source code project, primarily building/deploying.
Basing this whole thing on the idea that the workflow for using our code locally will revolve around doing a publishToMavenLocal and then generating a .yaml that uses the LOCAL version identifier in the GAV dependency field. Haven't decided if this stuff should link directly against compiled code via Gradle yet, but I'm leaning away from it because it'll introduce a ton of complexity.